[update][front] workspace.js 일부 profile.js로 분리
