+++ src/main/java/com/example/twogether/user/service/UserService.java
@@ -112,7 +112,7 @@ public class UserService {
     }
 
     @Transactional
-    public void defaultIcon(User user) throws IOException {
+    public void defaultIcon(User user) {
         try {
             user.editIcon("https://twogether.s3.ap-northeast-2.amazonaws.com/Icon/faed91e3-e029-45ee-a407-8efdfb178fce.png");
             userRepository.save(user);
